requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotli-python=1.0.9=py311h6a678d5_9
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.3.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.8=3
cycler=0.12.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py311h06a4308_0
fonttools=4.56.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py311h5eee18b_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.5=py311h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.8=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.13.0.11=0
libcurand=10.3.9.55=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgomp=14.2.0=h767d61c_2
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=14.2.0=h4852527_2
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=py311h5eee18b_0
matlabengine=24.2=pypi_0
matplotlib=3.10.1=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_2
mkl_fft=1.3.11=py311h5eee18b_0
mkl_random=1.2.8=py311ha02d727_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py311h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=py311h06a4308_0
numpy=2.0.1=py311h08b1b3b_1
numpy-base=2.0.1=py311hf175353_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.4.1=h7b32b05_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.1.0=py311hcea889d_0
pip=25.0.1=pypi_0
prettytable=3.15.1=pypi_0
protobuf=5.29.3=pypi_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.11=he870216_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.5.1=py3.11_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2025.1=pypi_0
pyyaml=6.0.2=py311h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.3=py311h06a4308_1
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.8.0=py311h06a4308_0
six=1.17.0=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=py311h06a4308_1
tbb=2021.8.0=hdb19cb5_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
torchaudio=2.5.1=py311_cu118
torchtriton=3.1.0=py311
torchvision=0.20.1=py311_cu118
tqdm=4.67.1=pypi_0
typing_extensions=4.12.2=py311h06a4308_0
tzdata=2025.1=pypi_0
urllib3=2.3.0=py311h06a4308_0
wcwidth=0.2.13=pypi_0
wheel=0.45.1=py311h06a4308_0
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
